+++ src/main/java/racinggame/view/MessageWriter.java
@@ -1,4 +1,4 @@
-package racinggame.ui;
+package racinggame.view;
 
 public interface MessageWriter {
     void write(final String message);
