+++ src/main/java/racinggame/view/RacingGameInputView.java
@@ -1,4 +1,4 @@
-package racinggame.ui;
+package racinggame.view;
 
 import racinggame.domain.RacingCar;
 import racinggame.domain.RacingCars;
