+++ src/test/java/racinggame/view/RacingGameInputViewTest.java
@@ -1,4 +1,4 @@
-package racinggame.ui;
+package racinggame.view;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
